+++ conf/settings.py
@@ -24,6 +24,7 @@ class RedisSetting(BaseModel):
     PORT: Optional[int] = 6379
     PASSWD: Optional[str] = ''
     USER: Optional[str] = ''
+    MAX_CONNECTIONS: Optional[int] = 10000
 
 
 class DBSetting(BaseModel):