+++ CHANGELOG.md
@@ -10,6 +10,7 @@ update/optimize
 
 - update celery config
 - update pagesize -> page_size
+- update Redis Connection Pool
 
 add